requirements.txt
google-adk==0.5.0
google-genai==1.14.0
google-api-python-client>=2.169.0
google-auth>=2.40.1
annotated-types
anyio
Authlib
cachetools
certifi
cffi
charset-normalizer
click
cryptography
Deprecated
docstring_parser
fastapi
google-api-core
google-auth-httplib2
google-auth-oauthlib
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-core
google-cloud-resource-manager
google-cloud-secret-manager
google-cloud-speech
google-cloud-storage
google-cloud-trace
google-crc32c
google-resumable-media
googleapis-common-protos
graphviz
grpc-google-iam-v1
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
httpx-sse
idna
importlib_metadata
mcp
numpy
oauthlib
opentelemetry-api
opentelemetry-exporter-gcp-trace
opentelemetry-resourcedetector-gcp
opentelemetry-sdk
opentelemetry-semantic-conventions
packaging
proto-plus
protobuf
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic-settings
pydantic_core
pyparsing
python-dateutil
python-dotenv
python-multipart
PyYAML
requests
requests-oauthlib
rsa
shapely
six
sniffio
SQLAlchemy
sse-starlette
starlette
typing-inspection
typing_extensions
tzlocal
uritemplate
urllib3
uvicorn
websockets
wrapt
zipp
docling
chromadb
langchain
langchain-community
langchain-google-genai
langchain-google-vertexai
langchain-huggingface
sentence-transformers
rapidfuzz
